Add support for wrap arounds in v_contain function
The v_contain function supports wrap arounds by subtracting 
(no_of_chars_per_row) * (no_of_rows) from the string pointer. This 
returns the pointer to the top of the relevant column and allows the 
while loop to continue going down the column. Note that words that are 
of the same length as the number of rows in the grid will loop back to 
the top.