Add support for wrap arounds in d_contain
The looping mechanism first waits for the for the position in the grid 
to reach the last column or the last row of the grid. Once this is true, 
no_of_chars_per_row + 1 is subtracted from the string pointer, which 
effectively moves the pointer up one row and left by one character. This 
is done until the top row or first column is reached.